+++ SMFL/LerpMoveTo.h
@@ -8,7 +8,7 @@ class LerpMoveTo : public IMoveTo
 public:
 	LerpMoveTo(float speed);
 	virtual ~LerpMoveTo() {};
-	virtual void MoveTo(GameObject& object, sf::Vector2f destination, double time);
+	virtual void MoveTo(GameObject& object, sf::Vector2f destination, float time);
 private:
 	float speed;
 };
